Create version 1.3.5-futrix3
Polygon click events coordinates are now relative to browser window

+++ googlemaps-demo/src/main/java/com/vaadin/tapio/googlemaps/demo/DemoUI.java
@@ -277,7 +277,7 @@ public class DemoUI extends UI {
 
         googleMap.addPolygonClickListener(new PolygonClickListener() {
             @Override
-            public void polygonClicked(GoogleMapPolygon clickedPolygon, double x, double y) {
+            public void polygonClicked(GoogleMapPolygon clickedPolygon, int x, int y) {
                 Notification.show("Clicked polygon " + clickedPolygon.getId() + " at " + x + ", " + y);
                 googleMap.removePolygonOverlay(clickedPolygon);
 //                addPolyOverlayButton.setEnabled(true);
